requirements.txt

Django>=3.1.0,<3.2.0
djangorestframework>=3.12.0,<3.13.0

python-dotenv>=0.14.0,<0.15.0

dj-database-url>=0.5.0,<0.6.0
psycopg2>=2.8.0,<2.9.0

djangorestframework-simplejwt>=4.4.0,<4.5.0
